% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automationFunctions.R
\name{Base_AutomationFunc_single}
\alias{Base_AutomationFunc_single}
\title{Print base automation functions for single or multi conditions}
\usage{
Base_AutomationFunc_single()
}
\value{
Invisible NULL. The function is called for its side effect of
deleting folders.
}
\description{
This function is used to print the mplus_automation code used in the
mplusParallel_automation function. This is done as convenience to ease custom
function generation.
}
\examples{
# Print the single condition function
Base_AutomationFunc_single()

# Print the multi-condition function
Base_AutomationFunc_multi()
}
\seealso{
\code{\link{mplusParallel_automation}} for the function that creates
these folders.
}
