% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeParFolders.R
\name{removeParFolders}
\alias{removeParFolders}
\title{Remove Parallel Processing Folders from mplusParallel_automation}
\usage{
removeParFolders(folder = "")
}
\arguments{
\item{folder}{The user-defined folder to search for parallel processing folders.
Should be the same as the one used for mplusParallel_automation.}
}
\value{
Invisible NULL. The function is called for its side effect of
deleting folders.
}
\description{
This function is used to delete all parallel processing folders (with names
containing the word "session") that were created by the \code{mplusParallel_automation}
function.
}
\examples{
\donttest{
# Assuming you have parallel processing folders in your current RStudio
# document's directory
removeParFolders(folder = 'user_defined_path')
}
}
\seealso{
\code{\link{mplusParallel_automation}} for the function that creates
these folders.
}
